Develop Pipeline for SFC using GED
In order for delay-embedding to work I need to downsample first. The whole pipeline is in LFPanal and the component can be visualized with visu_compSelecSFC. I have also experimented with using the trial segmented data instead of segmenting the continuous recording around each artefact (nan values in the LFP) and using method one of the GED (without delay-embedding and sphering). This produced "coupling" with 1/f+line noise which might be due to a faulty implementation because using continuous signal I could find simulated theta coupling (although within 1/f noise) in the data. I also tried out PPC as an alternative to GED