Adapt lfp2bundLFP for allSpks
The first part of lfp2bundLFP creates a structure with the ripple timestamps for each bundle.
A later part of the same script then does statistics on idx/ndx trials for ripples.

Also: Note that perm_ranksum does not work and initialize a variable for the permutation test for the number, length and density of ripples called perm_rppls that will be based on the mean or median.